refactor TellMorePage with context, update session with useSWR, api fields
